Only God sabi when we go finish this implementation

+++ core/src/main/java/com/cluster/core/data/model/CustomerValidationResponse.kt
@@ -44,7 +44,7 @@ class CustomerValidationResponse {
         var paymentReference: String? = null
 
         @SerialName("AmountDue")
-        var amountDue: Int? = null
+        var amountDue: Double? = null
 
         @SerialName("CustomerName")
         var customerName: String? = null
